+++ app/scrapers/sis_scraper.py
@@ -1,574 +1,121 @@
 import asyncio
 import json
-import os
-import re
 import time
 
 import aiohttp
 import bs4
-import lxml
 
-from app import logger
-
-from . import scraper_utils
-
-_OUTPUT_DATA_DIR = os.path.join(os.path.dirname(__file__), "scraper_data")
-_is_scraper_running = False
+# from app import logger
 
 
-async def get_subjects(session, term):
-    subjects_dict = {}
-    url = "https://sis.rpi.edu/rss/bwckctlg.p_display_courses"
-    url = f"{url}?term_in={term}&sel_crse_strt=&sel_crse_end=&sel_subj=&sel_levl=&sel_schd=&sel_coll=&sel_divs=&sel_dept=&sel_attr="
+async def get_subjects(
+    session: aiohttp.ClientSession, term: str
+) -> list[dict[str, str]]:
+    """
+    Fetches the list of subjects for a given term from the SIS API.
 
-    async with session.get(url) as response:
-        soup = bs4.BeautifulSoup(await response.text(), "lxml")
-
-        subjects = soup.find("select", {"name": "sel_subj"}).find_all("option")
-        for subject in subjects:
-            subject_code = subject.get("value")
-            subject_name = subject.get_text()
-            subjects_dict[subject_code] = subject_name
-
-        return subjects_dict
-
-
-async def get_courses(session, term, subject_code):
-    course_dict = {}
-    url = "https://sis.rpi.edu/rss/bwckctlg.p_display_courses"
-    params = (
-        f"term_in={term}&"
-        "call_proc_in=&"
-        "sel_subj=dummy&"
-        "sel_levl=dummy&"
-        "sel_schd=dummy&"
-        "sel_coll=dummy&"
-        "sel_divs=dummy&"
-        "sel_dept=dummy&"
-        "sel_attr=dummy&"
-        f"sel_subj={subject_code}&"
-        "sel_crse_strt=&"
-        "sel_crse_end=&"
-        "sel_title=&"
-        "sel_levl=%25&"
-        "sel_schd=%25&"
-        "sel_coll=%25&"
-        "sel_divs=%25&"
-        "sel_dept=%25&"
-        "sel_from_cred=&"
-        "sel_to_cred=&"
-        "sel_attr=%25"
-    )
-
-    url = f"{url}?{params}"
-    async with session.get(url) as response:
-        start = time.time()
-        soup = bs4.BeautifulSoup(await response.text(), "lxml")
-
-        tasks = []
-
-        subj_class_data = soup.find_all("td", class_="nttitle")
-        for class_link in subj_class_data:
-            class_link = class_link.find("a")
-            class_code_name = class_link.get_text().split(
-                " - "
-            )  # some class names include '-'
-            class_code, class_name = class_code_name[0], class_code_name[1]
-            subject_code, course_code = class_code.split(" ")
-
-            task = asyncio.create_task(
-                get_course_detail(session, term, subject_code, course_code)
-            )
-            tasks.append((task, class_code, class_name))
-
-        results = await asyncio.gather(*[task for task, _, _ in tasks])
-
-        for task_result, (_, class_code, class_name) in zip(results, tasks):
-            if task_result != None:
-                course_dict[class_code] = {
-                    "course_name": class_name,
-                    "course_detail": task_result,
-                }
-
-        end = time.time()
-        logger.debug(
-            f"Time taken to get term {term} -> Subject {subject_code}: {end - start:.2f} s"
-        )
-        return course_dict
-
-
-async def parse_prereqs(soup):
-    # prereq_pattern = r"(?:Prerequisite(?:s)?|Prerequisiste|Prerequisite or Corequisite):\s?(.*?)(?=\bCorequisite\b|$)"
-    prereq_label = soup.find("span", class_="fieldlabeltext", string="Prerequisites: ")
-
-    # No prerequisites found
-    if not prereq_label:
-        return []
-
-    prereq_info = []
-    for sibling in prereq_label.next_siblings:
-        if sibling.name != "br" and sibling.get_text(strip=True):
-            prereq_info.append(sibling.get_text(separator=" ", strip=True))
-    prereq_text = " ".join(prereq_info)
-
-    prereq_text_cleaned = re.sub(r"Undergraduate level ", "", prereq_text)
-    prereq_text_cleaned = re.sub(r"[()]", " ", prereq_text_cleaned)
-    prereq_text_cleaned = re.sub(r"\s+", " ", prereq_text_cleaned).strip()
-
-    individual_prereq = prereq_text_cleaned.split(" and ")
-
-    if individual_prereq == "":
-        return []
-    return individual_prereq
-
-
-async def fetch_CRNs(soup):
-    CRNs = []
-    section_title_wrappers = soup.find_all("th", class_="ddtitle")
-    for section_title_wrapper in section_title_wrappers:
-        section_title = section_title_wrapper.find("a").contents[0]
-        CRN = re.search(r"\d{5}", section_title).group()
-        CRNs.append(CRN)
-    return CRNs
-
-
-def get_seat_info(seat_table):
-    seat_info = seat_table.find_all("td")
-
-    capacity = seat_info[1].text
-    registered = seat_info[2].text
-    open_seats = seat_info[3].text
-
-    return capacity, registered, open_seats
-
-
-def get_section_crosslist(section_body):
-    # Cross List Courses
-    course_pattern = r"[A-Z]{4}\s\d{4}"
-    crosslist = []
-    if "Cross List Courses:" in section_body:
-        if "Prerequisites:" in section_body:
-            crosslistBody = section_body[
-                section_body.index("Cross List Courses:") : section_body.index(
-                    "Prerequisites"
-                )
-            ]
-        else:
-            crosslistBody = section_body[section_body.index("Cross List Courses:") :]
-        crosslist = re.findall(course_pattern, crosslistBody)
-
-    return crosslist
-
-
-def get_restrictions(section_body):
-    restriction_types = [
-        "Must be enrolled in one of the following Majors",
-        "May not be enrolled in one of the following Majors",
-        "Must be enrolled in one of the following Levels",
-        "May not be enrolled in one of the following Levels",
-        "Must be enrolled in one of the following Classifications",
-        "May not be enrolled as the following Classifications",
-    ]
-
-    major_restrictions = []
-    not_major_restrictions = []
-    level_restrictions = []
-    not_level_restrictions = []
-    classification_restrictions = []
-    not_classification_restrictions = []
-
-    # Remove excess characters
-    section_body = (
-        section_body.replace("&nbsp;", "").replace("\xa0", "").replace("   ", "")
-    )
-    while "\n\n" in section_body:
-        section_body = section_body.replace("\n\n", "\n")
-
-    if "Restrictions:" in section_body:
-        if "Prerequisites:" in section_body:
-            restriction_body = section_body[
-                section_body.index("Restrictions:") : section_body.index(
-                    "Prerequisites:"
-                )
-            ]
-        else:
-            restriction_body = section_body[section_body.index("Restrictions:") :]
-
-        for restriction in restriction_types:
-            match = re.search(
-                rf"{restriction}:\n([^\n]*(?:\n(?!Must be|May not be)[^\n]*)*)",
-                restriction_body,
-            )
-            if match:
-                restriction_data = match.group(1).strip().split("\n")
-
-                if restriction == restriction_types[0]:
-                    major_restrictions = restriction_data
-                elif restriction == restriction_types[1]:
-                    not_major_restrictions = restriction_data
-                elif restriction == restriction_types[2]:
-                    level_restrictions = restriction_data
-                elif restriction == restriction_types[3]:
-                    not_level_restrictions = restriction_data
-                elif restriction == restriction_types[4]:
-                    classification_restrictions = restriction_data
-                elif restriction == restriction_types[5]:
-                    not_classification_restrictions = restriction_data
-
-    return [
-        major_restrictions,
-        not_major_restrictions,
-        level_restrictions,
-        not_level_restrictions,
-        classification_restrictions,
-        not_classification_restrictions,
-    ]
-
-
-# Gets the restrictions, crosslist, capacity, registered, open_seats given a term and CRN
-async def get_section_data(session, term, CRN):
-    url = "https://sis.rpi.edu/rss/bwckschd.p_disp_detail_sched"
-    params = f"term_in={term}&crn_in={CRN}"
-    url = f"{url}?{params}"
-    async with session.get(url) as response:
-        soup = bs4.BeautifulSoup(await response.text(), "lxml")
-        seat_table = soup.find(
-            "table",
-            {
-                "class": "datadisplaytable",
-                "summary": "This layout table is used to present the seating numbers.",
-            },
-        )
-        section_body = soup.find("td", {"class": "dddefault"}).getText()
-
-        section_crosslist = get_section_crosslist(section_body)
-        restrictions = get_restrictions(section_body)
-        capacity, registered, open_seats = get_seat_info(seat_table)
-
-    return section_crosslist, restrictions, capacity, registered, open_seats
-
-
-async def get_section_info(session, term, soup):
-    CRNs = await fetch_CRNs(soup)
-
-    section_tables = soup.find_all(
-        "table",
+    Returned data format is as follows:
+    [
         {
-            "class": "datadisplaytable",
-            "summary": "This table lists the scheduled meeting times and assigned instructors for this class..",
+            "code": "ADMN",
+            "description": "Administrative Courses"
         },
-    )
+        ...
+    ]
+    """
+    url = "https://sis9.rpi.edu/StudentRegistrationSsb/ssb/classSearch/get_subject"
+    async with session.get(
+        url, params={"term": term, "offset": 1, "max": 100}
+    ) as response:
+        response.raise_for_status()
+        data = await response.json()
+        return data
 
-    sections_data = []
-    crosslist = []
-    major_restrictions, level_restrictions, classification_restrictions = [], [], []
-    not_major_restrictions, not_level_restrictions, not_classification_restrictions = (
-        [],
-        [],
-        [],
-    )
 
-    restrictions = {
-        "major": major_restrictions,
-        "not_major": not_major_restrictions,
-        "level": level_restrictions,
-        "not_level": not_level_restrictions,
-        "classification": classification_restrictions,
-        "not_classification": not_classification_restrictions,
+async def get_cookie(session: aiohttp.ClientSession, term: str) -> str:
+    """
+    Fetches a JSESSIONID cookie from the SIS API.
+
+    Must be re-called before each new request to get all courses in a subject.
+    """
+    url = "https://sis9.rpi.edu/StudentRegistrationSsb/ssb/term/search?mode=search"
+    async with session.get(
+        url,
+        params={"term": term},
+        headers={
+            "Accept": "text/html,application/xhtml+xml,application/xml;q=0.9,image/avif,image/webp,image/apng,*/*;q=0.8,application/signed-exchange;v=b3;q=0.7",
+            "Accept-Encoding": "gzip, deflate, br, zstd",
+            "Accept-Language": "en-US,en;q=0.9",
+            "Connection": "keep-alive",
+            "Host": "sis9.rpi.edu",
+            "Sec-Ch-Ua": '"Chromium";v="140", "Not=A?Brand";v="24", "Microsoft Edge";v="140"',
+            "Sec-Ch-Ua-Mobile": "?0",
+            "Sec-Ch-Ua-Platform": '"macOS"',
+            "Sec-Fetch-Dest": "document",
+            "Sec-Fetch-Mode": "navigate",
+            "Sec-Fetch-Site": "none",
+            "Sec-Fetch-User": "?1",
+            "Upgrade-Insecure-Requests": "1",
+        },
+    ) as response:
+        response.raise_for_status()
+        cookies = response.cookies
+        headers = response.headers
+        print(cookies or "No cookies")
+        print(headers)
+        return cookies["JSESSIONID"].value
+
+
+async def get_courses_in_subject(
+    session: aiohttp.ClientSession, jsessionid_cookie: str, term: str, subject: str
+) -> list[dict[str, str]]:
+    """
+    Fetches the list of courses for a given subject and term from the SIS API.
+
+    Returned data format is as follows:
+    [
+        {
+            "courseId": "12345",
+            "subject": "CSE",
+            "catalogNbr": "1010",
+            "titleLong": "Introduction to Computer Science",
+            ...
+        },
+        ...
+    ]
+    """
+    url = "https://sis9.rpi.edu/StudentRegistrationSsb/ssb/searchResults/searchResults"
+    params = {
+        "txt_subject": subject,
+        "txt_term": term,
+        "pageOffset": 0,
+        "pageMaxSize": 1000,
+        "sortColumn": "subjectDescription",
+        "sortDirection": "asc",
     }
+    headers = {"Cookie": f"JSESSIONID={jsessionid_cookie}"}
+    async with session.get(url, params=params, headers=headers) as response:
+        response.raise_for_status()
+        data = await response.json()
+        return data
 
-    count = 0
 
-    # per section
-    for section_table in section_tables:
-        schedule = {
-            "monday": {},
-            "tuesday": {},
-            "wednesday": {},
-            "thursday": {},
-            "friday": {},
-            "saturday": {},
-            "other": {},
+async def main():
+    async with aiohttp.ClientSession(
+        headers={
+            "User-Agent": "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/140.0.0.0 Safari/537.36 Edg/140.0.0.0",
+            "Accept": "*/*",
         }
-        instructors = []
-
-        # per row in section
-        for section_row in section_table.find_all("tr")[1:]:  # Skip the header row
-            # per cell in row
-            section_row_info = [
-                section_cells.text for section_cells in section_row.find_all("td")
-            ]
-
-            time = section_row_info[1]
-            days = section_row_info[2]
-            location = section_row_info[3]
-            if section_row_info[-1] != "TBA":
-                instructors_string = scraper_utils.clean_instructors(
-                    section_row_info[-1]
-                )
-                instructors = instructors_string.split(", ")
-
-            days = days.replace("\xa0", " ")
-
-            for day in scraper_utils.map_day_codes_to_days(days):
-                day_info = {"time": time, "location": location}
-                schedule[day] = day_info
-
-        CRN = CRNs[count]
-
-        section_crosslist, restrictions, capacity, registered, open_seats = (
-            await get_section_data(session, term, CRN)
-        )
-        # Crosslist
-        if section_crosslist != None:
-            for course in section_crosslist:
-                if course not in crosslist:
-                    crosslist.append(course)
-
-        section_major_restrictions = restrictions[0]
-        section_not_major_restrictions = restrictions[1]
-        section_level_restrictions = restrictions[2]
-        section_not_level_restrictions = restrictions[3]
-        section_classification_restrictions = restrictions[4]
-        section_not_classification_restrictions = restrictions[5]
-
-        # Restrictions
-        if section_major_restrictions != []:
-            for restriction in section_major_restrictions:
-                if restriction not in major_restrictions:
-                    major_restrictions.append(restriction)
-
-        if section_not_major_restrictions != []:
-            for restriction in section_not_major_restrictions:
-                if restriction not in not_major_restrictions:
-                    not_major_restrictions.append(restriction)
-
-        if section_level_restrictions != []:
-            for restriction in section_level_restrictions:
-                if restriction not in level_restrictions:
-                    level_restrictions.append(restriction)
-        if section_not_level_restrictions != []:
-            for restriction in section_not_level_restrictions:
-                if restriction not in not_level_restrictions:
-                    not_level_restrictions.append(restriction)
-
-        if section_classification_restrictions != []:
-            for restriction in section_classification_restrictions:
-                if restriction not in classification_restrictions:
-                    classification_restrictions.append(restriction)
-        if section_not_classification_restrictions != []:
-            for restriction in section_not_classification_restrictions:
-                if restriction not in not_classification_restrictions:
-                    not_classification_restrictions.append(restriction)
-
-        restrictions = {
-            "major": major_restrictions,
-            "not_major": not_major_restrictions,
-            "level": level_restrictions,
-            "not_level": not_level_restrictions,
-            "classification": classification_restrictions,
-            "not_classification": not_classification_restrictions,
-        }
-
-        section_entry = {
-            "CRN": CRN,
-            "instructor": instructors,
-            "schedule": schedule,
-            "capacity": int(capacity),
-            "registered": int(registered),
-            "open": int(open_seats),
-        }
-
-        sections_data.append(section_entry)
-        count += 1
-
-    return crosslist, restrictions, sections_data
-
-
-async def fetch_crosslist_restriction_section(session, term, subject_code, course_code):
-    url = "https://sis.rpi.edu/rss/bwckctlg.p_disp_listcrse"
-    param = f"term_in={term}&subj_in={subject_code}&crse_in={course_code}&schd_in=L"
-    url = f"{url}?{param}"
-    async with session.get(url) as response:
-        soup = bs4.BeautifulSoup(await response.text(), "lxml")
-
-        is_section_found = soup.find("caption", class_="captiontext")
-        if is_section_found == None:
-            return None, None, None
-
-        response_crosslist, response_restrictions, sections_data = (
-            await get_section_info(session, term, soup)
-        )
-
-        return response_crosslist, response_restrictions, sections_data
-
-
-async def parse_attributes(soup, term, subject_code, course_code):
-    attributes = []
-    attribute_label = soup.find(
-        "span", class_="fieldlabeltext", string="Course Attributes: "
-    )
-
-    if not attribute_label:
-        return attributes
-
-    attributes = attribute_label.next.next.next.strip().split(", ")
-
-    return attributes
-
-
-async def parse_crosslist(soup):
-    soup.find("span", class_="fieldlabeltext", string="Cross Listed: ")
-
-
-async def get_course_detail(session, term, subject_code, course_code):
-    url = "https://sis.rpi.edu/rss/bwckctlg.p_disp_course_detail"
-    params = (
-        f"cat_term_in={term}&subj_code_in={subject_code}&crse_numb_in={course_code}"
-    )
-    url = f"{url}?{params}"
-    info = {
-        "description": "",
-        "corequisite": [],
-        "prerequisite": [],
-        "crosslist": [],
-        "attributes": [],
-        "restrictions": {"major": [], "level": [], "classification": [], "grade": []},
-        "credits": {"min": 0, "max": 0},
-        "offered": "",
-    }
-
-    async with session.get(url) as response:
-        soup = bs4.BeautifulSoup(await response.text(), "lxml")
-
-        course_data = (
-            soup.find("td", class_="ntdefault").contents[0].strip().split("\n")
-        )
-
-        crosslist, restrictions, sections_data = (
-            await fetch_crosslist_restriction_section(
-                session, term, subject_code, course_code
-            )
-        )
-
-        if sections_data == None:
-            return None
-
-        coreq_pattern = r"Corequisite:\s?(.*)"
-
-        for _ in course_data:
-            if _ != "":
-                if info["description"] == "":  # if not empty
-                    info["description"] = _.strip()
-                else:
-                    if "Corequisite:" in _:
-                        try:
-                            if "Corequisite:" in _:
-                                coreq_match = re.search(coreq_pattern, _, re.IGNORECASE)
-                                info["corequisite"] = [coreq_match.group(1).strip()]
-                        except OSError as e:
-                            logger.error(e)
-                            logger.error(f"ERROR Coreq: {subject_code} - {course_code}")
-                            info["corequisite"] = None
-                    elif "Prerequisite:" in _:
-                        try:
-                            info["prerequisite"] = await parse_prereqs(soup)
-                        except OSError as e:
-                            logger.error(e)
-                            logger.error(
-                                f"ERROR Prereq: {subject_code} - {course_code}"
-                            )
-                            info["prerequisite"] = None
-                    elif "Credit Hours:" in _:
-                        try:
-                            min_max = scraper_utils.get_min_max(_.split(":")[1].strip())
-                            info["credits"] = {"min": min_max[0], "max": min_max[1]}
-                        except OSError as e:
-                            logger.error(e)
-                            logger.error(
-                                f"ERROR Credit: {subject_code} - {course_code}"
-                            )
-                            info["credits"] = None
-                    elif "When Offered: " in _:
-                        try:
-                            info["offered"] = _.split(":")[1].strip()
-                        except OSError as e:
-                            logger.error(e)
-                            logger.error(
-                                f"ERROR Offered: {subject_code} - {course_code}"
-                            )
-                            info["offered"] = None
-                    elif "Cross Listed:" in _:
-                        try:
-                            info["crosslist"] = [_.split(":")[1].strip()]
-                        except OSError as e:
-                            logger.error(e)
-                            logger.error(
-                                f"ERROR crosslist: {subject_code} - {course_code}"
-                            )
-                            info["crosslist"] = None
-
-        info["crosslist"] = crosslist
-        info["attributes"] = await parse_attributes(
-            soup, term, subject_code, course_code
-        )
-        info["restrictions"] = restrictions
-        info["sections"] = sections_data
-
-    return info
-
-
-async def main() -> None:
-    global _is_scraper_running
-    if _is_scraper_running:
-        logger.info("Cannot run scraper as it is already running")
-        return False
-    _is_scraper_running = True
-    if not os.path.exists(_OUTPUT_DATA_DIR):
-        os.makedirs(_OUTPUT_DATA_DIR)
-    total_start = time.time()
-    start_year = 2024
-    end_year = 2025
-    logger.info(f"Starting SIS course scraper for {start_year} - {end_year}")
-    for i in range(start_year, end_year + 1):
-        for semester in ["fall", "spring", "summer"]:
-            term = scraper_utils.get_term(i, semester)
-            logger.info(f"Scraping term {term}")
-            async with aiohttp.ClientSession() as session:
-                start = time.time()
-                subjects = await get_subjects(session, term=term)
-                all_courses = {}
-
-                tasks = []
-
-                subject_metadata = []
-                for subject_code, subject_name in subjects.items():
-                    task = asyncio.create_task(
-                        get_courses(
-                            session=session, term=term, subject_code=subject_code
-                        )
-                    )
-                    tasks.append(task)
-                    subject_metadata.append((subject_code, subject_name))
-
-                courses_by_subject = await asyncio.gather(*tasks)
-                for (subject_code, subject_name), courses in zip(
-                    subject_metadata, courses_by_subject
-                ):
-                    all_courses[subject_code] = {
-                        "subject_name": subject_name,
-                        "courses": courses,
-                    }
-                with open(f"{_OUTPUT_DATA_DIR}/{term}.json", "w") as f:
-                    json.dump(all_courses, f, indent=4)
-                end = time.time()
-                logger.info(f"Time elapsed for term {term}: {end - start:.2f} s")
-
-    total_end = time.time()
-    logger.info(f"Total elapsed time: {total_end - total_start:.2f} s")
-    _is_scraper_running = False
-    return True
+    ) as session:
+        term = "202409"
+        subjects = await get_subjects(session, term)
+        cookie = await get_cookie(session, term)
+        data = await get_courses_in_subject(session, cookie, term, subjects[0]["code"])
+        print(json.dumps(data, indent=4))
 
 
 if __name__ == "__main__":
+    start_time = time.time()
     asyncio.run(main())
+    end_time = time.time()
+    # logger.info(f"Execution time: {end_time - start_time} seconds")
